+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## New features
 
-  * Time limit for tasks (option ``--time-limit``)
+  * Time limit and Time request for tasks (options ``--time-limit`` and ``--time-request``)
+  * Time limit for workers
   * Job and task times are shown in job information tables
   * Integers in command line options can be now written with an underscore separator (e.g. ``--array=1-1_000``)
   * The default path of stdout and stderr has been changed to Storing tasks default stdout by new rule `` job-%{JOB_ID}/[stdout/stderr].%{TASK_ID}``